Merged in upstream user config file search logic
This update brings the user config file search logic closer to what's done in the upstream `htop` (v1.0.2).

In upstream [revision 13][1], the `$HTOPRC` environmental variable was introduced as a way to point to an alternate config file, but a typo had it looking for `$HOMERC` instead. This was fixed in [revision 275][2], but is still in `htop-osx`.

This commit also incorporates the default config file path change from `~/.htoprc` to the [XDG Base Directory Specification][3] compliant path, `~/.config/htop/htoprc`.

Before:

1) `$HOMERC`
2) `$HOME_ETC/.htoprc`
3) `$HOME/.htoprc`

After:

1) `$HTOPRC`
2) `$XDG_CONFIG_HOME/htop/htoprc`
3) `$HOME/.config/htop/htoprc`
4) `$HOME/.htoprc` (legacy)

If the legacy config file is found, it's migrated to the new default location.

For reference, here's the upstream commit log for [Settings.c][4].

[1]: http://htop.svn.sourceforge.net/viewvc/htop/trunk/Settings.c?revision=13&view=markup
[2]: http://htop.svn.sourceforge.net/viewvc/htop/trunk/Settings.c?r1=267&r2=275
[3]: http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html
[4]: http://htop.svn.sourceforge.net/viewvc/htop/trunk/Settings.c?view=log