Initial work for Fan RPM monitoring.
Right now it collapses them together like CPU_TEMP does (see issue #5), and has hard coded values for the "heat" of the fan speed. What it should do instead is to use the current speed vs. the max speed (which I'm already pulling) and have the heat be based off of a percentage towards that. Say "warm" is 50%, "hot" is 80%, etc. Ideally this would also support other meter types than just text, as a fan RPM range is fairly easy to visualize-- I'm just not sure how to do other meter types yet.

This also does not currently track when the fan speed is "cold", ie, below the minimum value. I'm not sure how to represent that yet.